NMS-8043: Code cleanup
