+++ features/topology-map/plugins/org.opennms.features.topology.plugins.netutils/src/test/java/org/opennms/features/topology/netutils/internal/DummyUI.java
@@ -1,0 +1,76 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2016 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2016 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.features.topology.netutils.internal;
+
+import java.net.URI;
+import java.net.URISyntaxException;
+
+import org.easymock.EasyMock;
+import org.easymock.IAnswer;
+
+import com.vaadin.server.Page;
+import com.vaadin.server.VaadinRequest;
+import com.vaadin.ui.Notification;
+import com.vaadin.ui.UI;
+
+// Empty Application
+class DummyUI extends UI {
+    private static final long serialVersionUID = -6761162156810032609L;
+
+    private boolean notified = false;
+
+    @Override
+    public void init(VaadinRequest request) {}
+
+    @Override
+    public Page getPage() {
+        Page page = EasyMock.createMock(Page.class);
+        try {
+            EasyMock.expect(page.getLocation()).andReturn(new URI("http://localhost:8080/servlet/")).anyTimes();
+            page.showNotification(EasyMock.anyObject(Notification.class));
+            // If Notification.show() is called, then set didNotify to true
+            EasyMock.expectLastCall().andAnswer(new IAnswer<Object>() {
+                @Override
+                public Object answer() throws Throwable {
+                    System.out.println("Notification was called: " + ((Notification)EasyMock.getCurrentArguments()[0]).getCaption());
+                    notified = true;
+                    return null;
+                }
+            }).anyTimes();
+        } catch (URISyntaxException e) {
+            // Should never be thrown
+        }
+        EasyMock.replay(page);
+        return page;
+    }
+
+    public boolean isNotified() {
+        return notified;
+    }
+}
